Internal Cargo and other changes
Removed or disabled the simulated sling loading (WIP to put it back and make it compatible with these changes, or optionally can remove simulated sling loading) and modified it to be internal cargo loading

Crates:
Come in both internal and external-only. Internal cargos can still be carried via sling, but can also be loaded internally by using the Other > CTLD > CTLD Commands > Load Crate menu item. Note that the cargo doors have to be close to the load, and open. So in the case of the Huey this is at the 2-4 o'clock position, or at the 8-10 o'clock position. For the Mi-8 this is at the rear (note: you do not need to have it under the tail boom to load, it can be a few metres back).

Unloading is a similar process, with the Other > CTLD > CTLD Commands > Drop Crate menu item. Note the cargo doors must be open. In the Huey, if the right, or both doors are open, the cargo will unload to the right. If the left door only is open, it will unload to the left.

You cannot carry internal cargo when you are carrying troops. You can however simultaneously carry internal, and external cargo. Unfortunately due to DCS restrictions, adding weight for internal cargo is apparently still unsupported, so it will have no effect on helicopter performance, unlike external slung loads.

JTAC has been upgraded to optionally use the separate AutoLase script if it is already in use on the server.